From: Alexander Petry <alexander.petry@itwm.fraunhofer.de>

git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@1405 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/sdpa/trunk/sdpa/daemon/WorkerManager.hpp
@@ -1,3 +1,20 @@
+/*
+ * =====================================================================================
+ *
+ *       Filename:  WorkerManager.hpp
+ *
+ *    Description:  Worker manager
+ *
+ *        Version:  1.0
+ *        Created:
+ *       Revision:  none
+ *       Compiler:  gcc
+ *
+ *         Author:  Dr. Tiberiu Rotaru, tiberiu.rotaru@itwm.fraunhofer.de
+ *        Company:  Fraunhofer ITWM
+ *
+ * =====================================================================================
+ */
 #ifndef SDPA_DAEMON_WORKER_MANAGER_HPP
 #define SDPA_DAEMON_WORKER_MANAGER_HPP 1
 
@@ -20,7 +37,7 @@ namespace sdpa { namespace daemon {
 	  Worker::ptr_t &findWorker(const Worker::worker_id_t& worker_id) throw(WorkerNotFoundException);
 	  void addWorker(const Worker::ptr_t &pWorker);
 	  Worker::ptr_t &getNextWorker() throw (NoWorkerFoundException);
-	  int numberOfWorkers() { return worker_map_.size(); }
+	  size_t numberOfWorkers() { return worker_map_.size(); }
 
 	  //only for testing purposes!
 	  friend class sdpa::tests::DaemonFSMTest_SMC;
